requirements.txt
psycopg2-binary
folium
pandas
geopy
requests
g4f